requirements.txt
asgiref==3.10.0
Brotli==1.1.0
cffi==2.0.0
channels==4.3.1
cssselect2==0.8.0
Django==5.2.7
django-storages==1.14.6
djangorestframework==3.16.1
drf-yasg==1.21.11
fonttools==4.60.1
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.2.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
